test/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
ta==0.10.2
ccxt==4.1.77
asyncio==3.4.3
aiohttp==3.9.1
schedule==1.2.0
psutil==5.9.6
plotly==5.17.0
websocket-client==1.6.4
python-binance==1.0.19
yfinance==0.2.18
scikit-learn==1.3.2
